+++ _Resources/user-service/lib/UserService.js
